requirements.txt
asgiref==3.2.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
pylint==2.4.4
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
selenium==3.141.0
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
wrapt==1.11.2
django-heroku
gunicorn
django-crispy-forms
whitenoise